+++ lib/src/app/repo/external_repo/external_repo_exporter.dart
@@ -4,3 +4,4 @@ library external_repo_exporter;
 export 'abstract_external_repo.dart';
 export 'crud_repository_external.dart';
 export 'crud_repository_external_async.dart';
+export 'json_converter.dart';